polishing the Si5345 python
