+++ python/surf/devices/silabs/__init__.py
@@ -8,4 +8,6 @@
 ## may be copied, modified, propagated, or distributed except according to 
 ## the terms contained in the LICENSE.txt file.
 ##############################################################################
-from surf.devices.silabs._Si5345 import *
+from surf.devices.silabs._Si5345Pages import *
+from surf.devices.silabs._Si5345Lite  import *
+from surf.devices.silabs._Si5345      import *
